backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
more-itertools==10.7.0
psycopg==3.2.9
python-dotenv==1.1.1
sqlacodegen==3.0.0
SQLAlchemy==2.0.41
typeguard==4.4.4
typing_extensions==4.14.0
Werkzeug==3.1.3
